DM-9803:   Split the LSSTMultiObjectSearch.java to two files, LSSTMultiObjectSearch.java and LSSTConcurrentSearch.java.  The LSSTConcurrentSearch.java can be re-used.
